refactor: Address PR review feedback - type conversion and file organization
- Change Int64 to Int for maxFailures and minPopularity configuration parameters
- Split ConfigSource enum into separate file from ConfigurationError.swift
- Split ConfigurationKeys enum into separate file from ConfigurationLoader.swift
- Simplify readInt() helper by removing unnecessary Int64 conversion
- Make ConfigurationKeys internal instead of private for module-wide reusability

This addresses PR #22 review comments requesting more idiomatic Swift types
and following the one-type-per-file guideline.

+++ Sources/CelestraCloudKit/Services/CloudKitService+Celestra.swift
@@ -73,7 +73,7 @@ extension CloudKitService {
   /// Query feeds with optional filters (demonstrates QueryFilter and QuerySort)
   public func queryFeeds(
     lastAttemptedBefore: Date? = nil,
-    minPopularity: Int64? = nil,
+    minPopularity: Int? = nil,
     limit: Int = 100
   ) async throws -> [Feed] {
     var filters: [QueryFilter] = []
@@ -85,7 +85,7 @@ extension CloudKitService {
 
     // Filter by minimum popularity if provided
     if let minPop = minPopularity {
-      filters.append(.greaterThanOrEquals("subscriberCount", .int64(Int(minPop))))
+      filters.append(.greaterThanOrEquals("subscriberCount", .int64(minPop)))
     }
 
     // Query with filters and sort by feedURL (always queryable+sortable)
